+++ crates/icu/src/utils/mod.rs
@@ -2,3 +2,4 @@ pub mod instructions;
 pub mod rand;
 pub mod time;
 pub mod wasm;
+pub mod serialization;