+++ crates/icu/src/utils/serialization.rs
@@ -1,17 +1,9 @@
 use minicbor_serde::{from_slice, to_vec};
-use serde::{Serialize, de::DeserializeOwned};
+use serde::{de::DeserializeOwned, Serialize};
 use std::fmt::Debug;
 use thiserror::Error as ThisError;
 
-///
-/// Serialize/Deserialize
-/// for consistent use of minicbor (super efficient, no_std)
-///
-
-///
-/// SerializeError
-///
-
+/// Serialize/Deserialize via MiniCBOR for efficient, no_std-friendly codecs.
 #[derive(Debug, ThisError)]
 pub enum SerializeError {
     #[error("serialize error: {0}")]
@@ -21,22 +13,19 @@ pub enum SerializeError {
     Deserialize(String),
 }
 
-// serialize
 pub fn serialize<T>(t: &T) -> Result<Vec<u8>, SerializeError>
 where
     T: Serialize,
 {
     let bytes = to_vec(t).map_err(|e| SerializeError::Serialize(e.to_string()))?;
-
     Ok(bytes)
 }
 
-// deserialize
 pub fn deserialize<T>(bytes: &[u8]) -> Result<T, SerializeError>
 where
     T: DeserializeOwned,
 {
     let t: T = from_slice(bytes).map_err(|e| SerializeError::Deserialize(e.to_string()))?;
-
     Ok(t)
 }
+